Refactor into platform class.
